+++ layout/embed/map/component.jsx
@@ -1,0 +1,542 @@
+import React, {
+  useState,
+  useCallback,
+  useEffect,
+} from 'react';
+import PropTypes from 'prop-types';
+import classnames from 'classnames';
+import isEmpty from 'lodash/isEmpty';
+import { useDebouncedCallback } from 'use-debounce';
+import { useRouter } from 'next/router';
+import { Popup } from 'react-map-gl';
+import {
+  Legend,
+  LegendListItem,
+  LegendItemTypes,
+} from 'vizzuality-components';
+import Link from 'next/link';
+import { CancelToken } from 'axios';
+
+// components
+import LayoutEmbed from 'layout/layout/layout-embed';
+import Map from 'components/map';
+import LayerManager from 'components/map/layer-manager';
+import MapControls from 'components/map/controls';
+import ZoomControls from 'components/map/controls/zoom';
+import ResetViewControls from 'components/map/controls/reset-view';
+import LayerPopup from 'components/map/popup';
+import Spinner from 'components/ui/Spinner';
+import Icon from 'components/ui/icon';
+
+// hooks
+import useIsFavorite from 'hooks/favorite/is-favorite';
+
+// services
+import {
+  createFavourite,
+  deleteFavourite,
+} from 'services/favourites';
+import {
+  fetchArea,
+} from 'services/areas';
+import {
+  fetchGeostore,
+} from 'services/geostore';
+
+// constants
+import { DEFAULT_VIEWPORT, MAPSTYLES } from 'components/map/constants';
+
+// utils
+import { paramIsTrue } from 'utils/utils';
+import { isLoadedExternally } from 'utils/embed';
+import { getUserAreaLayer } from 'components/ui/map/utils';
+
+const LayoutEmbedMap = (props) => {
+  const {
+    widget,
+    isLoading,
+    isError,
+    layerGroups,
+    user,
+    webshot,
+    mapProps: {
+      basemap,
+      labels,
+      boundaries,
+      bounds,
+    },
+    activeLayers,
+    activeInteractiveLayers,
+  } = props;
+  const {
+    query: {
+      disableZoom,
+      legendExpanded,
+      areas,
+    },
+  } = useRouter();
+  const [mapState, setMapState] = useState({
+    viewport: DEFAULT_VIEWPORT,
+    interaction: {
+      data: {},
+      selected: 0,
+      point: null,
+    },
+    bounds,
+  });
+  const [displayedLayers, setDisplayedLayers] = useState([
+    ...activeLayers,
+  ]);
+  const [modalVisibility, setModalVisibility] = useState();
+
+  const {
+    name,
+    description,
+    dataset,
+    id,
+    thumbnailUrl,
+  } = widget || {};
+  const {
+    isFavorite,
+    data: favorite,
+    refetch: refetchFavorites,
+  } = useIsFavorite(widget?.id, user?.token);
+
+  const handleModalVisibility = useCallback(() => {
+    const toggleVisibility = !modalVisibility;
+    setModalVisibility(toggleVisibility);
+  }, [modalVisibility]);
+
+  const toggleFavorite = useCallback(async () => {
+    const toggleIsFavorite = !isFavorite;
+    if (toggleIsFavorite) {
+      try {
+        await createFavourite(user.token, {
+          resourceType: 'widget',
+          resourceId: id,
+        });
+        refetchFavorites();
+      } catch (e) {
+        // do something
+      }
+    } else {
+      try {
+        await deleteFavourite(user.token, favorite.id);
+        refetchFavorites();
+      } catch (e) {
+        // do something
+      }
+    }
+  }, [isFavorite, id, favorite, user, refetchFavorites]);
+
+  const onChangeInteractiveLayer = useCallback((selected) => {
+    setMapState((prevMapState) => ({
+      ...prevMapState,
+      interaction: {
+        ...prevMapState.interaction,
+        selected,
+      },
+    }));
+  }, []);
+
+  const onClickLayer = useCallback(({ features, lngLat }) => {
+    const { interaction } = mapState;
+    let interactions = {};
+
+    // if the user clicks on a zone where there is no data in any current layer
+    // we will reset the current interaction of those layers to display "no data available" message
+    if (!features.length) {
+      interactions = Object.keys(interaction.data).reduce((accumulator, currentValue) => ({
+        ...accumulator,
+        [currentValue]: {},
+      }), {});
+    } else {
+      interactions = features.reduce((accumulator, currentValue) => ({
+        ...accumulator,
+        [currentValue.layer.source]: { data: currentValue.properties },
+      }), {});
+    }
+
+    setMapState((prevMapState) => ({
+      ...prevMapState,
+      interaction: {
+        ...prevMapState.interaction,
+        data: interactions,
+        point: {
+          longitude: lngLat[0],
+          latitude: lngLat[1],
+        },
+      },
+    }));
+  }, [mapState]);
+
+  const [handleViewport] = useDebouncedCallback((viewport) => {
+    setMapState((prevMapState) => ({
+      ...prevMapState,
+      viewport,
+    }));
+  }, 250);
+
+  const handleZoom = useCallback((zoom) => {
+    setMapState((prevMapState) => ({
+      ...prevMapState,
+      viewport: {
+        ...prevMapState.viewport,
+        zoom,
+        // transitionDuration is always set to avoid mixing
+        // durations of other actions (like flying)
+        transitionDuration: 250,
+      },
+    }));
+  }, []);
+
+  const handleResetView = useCallback(() => {
+    setMapState((prevMapState) => ({
+      ...prevMapState,
+      viewport: {
+        bearing: 0,
+        pitch: 0,
+        // transitionDuration is always set to avoid mixing
+        // durations of other actions (like flying)
+        transitionDuration: 250,
+      },
+    }));
+  }, []);
+
+  const handleClosePopup = useCallback(() => {
+    setMapState((prevMapState) => ({
+      ...prevMapState,
+      interaction: {
+        data: {},
+        selected: 0,
+        point: null,
+      },
+    }));
+  }, []);
+
+  useEffect(() => {
+    setDisplayedLayers((prevLayers) => [
+      ...prevLayers.filter(({ provider }) => provider === 'geojson'),
+      ...activeLayers,
+    ]);
+  }, [activeLayers]);
+
+  useEffect(() => {
+    setMapState((prevMapState) => ({
+      ...prevMapState,
+      bounds,
+    }));
+  }, [bounds]);
+
+  useEffect(() => {
+    const cancelToken = CancelToken.source();
+
+    const loadAreas = async () => {
+      try {
+        const areasId = areas.split(',');
+        const areasData = await Promise.all(
+          areasId.map((_areaId) => fetchArea(_areaId, {}, {
+            cancelToken: cancelToken.token,
+            Authorization: user.token,
+          })),
+        );
+        const geostores = await Promise.all(
+          areasData.map(
+            ({ geostore }) => fetchGeostore(geostore, { cancelToken: cancelToken.token }),
+          ),
+        );
+        const areaLayers = geostores.map(({ id: geostoreId, geojson }, index) => getUserAreaLayer({ id: `${geostoreId}-${index}`, geojson }));
+
+        setDisplayedLayers((prevLayers) => [
+          ...areaLayers,
+          ...prevLayers.filter(({ provider }) => provider !== 'geojson'),
+        ]);
+
+        if (geostores.length === 1) {
+          setMapState((prevMapState) => ({
+            ...prevMapState,
+            bounds: {
+              bbox: geostores[0].bbox,
+              options: {
+                padding: 50,
+              },
+            },
+          }));
+        }
+      } catch (e) {
+        //  do something
+      }
+    };
+
+    if (user.token) loadAreas();
+
+    return () => { cancelToken.cancel('Fetching geostore: operation canceled by the user.'); };
+  }, [areas, user]);
+
+  const {
+    viewport,
+    interaction,
+    bounds: mapBounds,
+  } = mapState;
+  const { pitch, bearing } = viewport;
+  const favoriteIcon = isFavorite ? 'star-full' : 'star-empty';
+  const isExternal = typeof window !== 'undefined' ? isLoadedExternally(window?.location?.href) : false;
+  const resetViewBtnClass = classnames({
+    '-with-transition': true,
+    '-visible': pitch !== 0 || bearing !== 0,
+  });
+
+  return (
+    <>
+      {isLoading && (
+        <LayoutEmbed
+          title="Loading widget..."
+          description=""
+        >
+          <div className="c-embed-widget -map">
+            <Spinner
+              isLoading={isLoading}
+              className="-light"
+            />
+          </div>
+        </LayoutEmbed>
+      )}
+
+      {isError && (
+        <LayoutEmbed
+          title="Resource Watch"
+          description=""
+        >
+          <div className="c-embed-widget -map">
+            <div className="widget-title">
+              <h4>–</h4>
+            </div>
+            <div className="widget-content">
+              <p>{'Sorry, the widget couldn\'t be loaded'}</p>
+            </div>
+
+            {isExternal && (
+              <div className="widget-footer">
+                <Link
+                  href="/"
+                >
+                  <a
+                    href="/"
+                    target="_blank"
+                    rel="noopener noreferrer"
+                  >
+                    <img
+                      className="embed-logo"
+                      src="/static/images/logo-embed.png"
+                      alt="Resource Watch"
+                    />
+                  </a>
+                </Link>
+              </div>
+            )}
+          </div>
+        </LayoutEmbed>
+      )}
+
+      {!isLoading && (
+        <LayoutEmbed
+          title={name}
+          description={`${description || ''}`}
+          {...thumbnailUrl && { thumbnailUrl }}
+        >
+          <div className="c-embed-widget -map">
+            {!webshot && (
+              <div className="widget-title">
+                <Link
+                  href={`/explore/${dataset}`}
+                >
+                  <a
+                    href={`/explore/${dataset}`}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                  >
+                    <h4>{name}</h4>
+                  </a>
+                </Link>
+                <div className="buttons">
+                  {
+                    (user?.id) && (
+                      <button
+                        type="button"
+                        onClick={toggleFavorite}
+                      >
+                        <Icon name={`icon-${favoriteIcon}`} className="c-icon -small" />
+                      </button>
+                    )
+                  }
+                  <button
+                    type="button"
+                    aria-label={`${modalVisibility ? 'Close' : 'Open'} information modal`}
+                    onClick={handleModalVisibility}
+                  >
+                    <Icon name={`icon-${modalVisibility ? 'cross' : 'info'}`} className="c-icon -small" />
+                  </button>
+                </div>
+              </div>
+            )}
+
+            <div className={classnames('widget-content', { '-external': isExternal })}>
+              <Map
+                mapboxApiAccessToken={process.env.RW_MAPBOX_API_TOKEN}
+                onClick={onClickLayer}
+                interactiveLayerIds={activeInteractiveLayers}
+                mapStyle={MAPSTYLES}
+                viewport={viewport}
+                basemap={basemap}
+                labels={labels}
+                bounds={mapBounds}
+                fitBoundsOptions={{ transitionDuration: 0 }}
+                boundaries={boundaries}
+                scrollZoom={false}
+                onViewportChange={handleViewport}
+              >
+                {(_map) => (
+                  <>
+                    <LayerManager
+                      map={_map}
+                      layers={displayedLayers}
+                    />
+
+                    {!isEmpty(interaction.point)
+                      && activeLayers.length
+                      && !isEmpty(interaction.data) && (
+                        <Popup
+                          {...interaction.point}
+                          closeButton
+                          closeOnClick={false}
+                          onClose={handleClosePopup}
+                          className="rw-popup-layer"
+                          maxWidth="250px"
+                        >
+                          <LayerPopup
+                            data={{
+                              // data available in certain point
+                              layersInteraction: interaction.data,
+                              // ID of the layer will display data (defaults into the first layer)
+                              layersInteractionSelected: interaction.selected,
+                              // current active layers to get their layerConfig attributes
+                              layers: activeLayers,
+                            }}
+                            latlng={{
+                              lat: interaction.point.latitude,
+                              lng: interaction.point.longitude,
+                            }}
+                            onChangeInteractiveLayer={onChangeInteractiveLayer}
+                          />
+                        </Popup>
+                    )}
+                  </>
+                )}
+              </Map>
+
+              {!webshot && (
+                <MapControls>
+                  {!paramIsTrue(disableZoom) && (
+                    <ZoomControls
+                      viewport={viewport}
+                      onClick={handleZoom}
+                    />
+                  )}
+                  <ResetViewControls
+                    className={resetViewBtnClass}
+                    onResetView={handleResetView}
+                  />
+                </MapControls>
+              )}
+
+              <div className="c-legend-map -embed">
+                <Legend
+                  maxHeight={200}
+                  sortable={false}
+                  expanded={paramIsTrue(!!legendExpanded)}
+                >
+                  {layerGroups.map((lg, i) => (
+                    <LegendListItem
+                      index={i}
+                      key={lg.dataset}
+                      layerGroup={lg}
+                    >
+                      <LegendItemTypes />
+                    </LegendListItem>
+                  ))}
+                </Legend>
+              </div>
+              {modalVisibility && (
+                <div className="widget-modal">
+                  {!description && (
+                    <p>No additional information is available</p>
+                  )}
+                  {description && (
+                    <>
+                      <h4>Description</h4>
+                      <p>{description}</p>
+                    </>
+                  )}
+                </div>
+              )}
+            </div>
+            {(isExternal && !webshot) && (
+              <div className="widget-footer -map">
+                Powered by
+                <Link
+                  href="/"
+                >
+                  <a
+                    href="/"
+                    target="_blank"
+                    rel="noopener noreferrer"
+                  >
+                    <img
+                      className="embed-logo"
+                      src="/static/images/logo-embed.png"
+                      alt="Resource Watch"
+                    />
+                  </a>
+                </Link>
+              </div>
+            )}
+          </div>
+        </LayoutEmbed>
+      )}
+    </>
+  );
+};
+
+LayoutEmbedMap.propTypes = {
+  widget: PropTypes.shape({
+    id: PropTypes.string,
+    name: PropTypes.string,
+    description: PropTypes.string,
+    dataset: PropTypes.string,
+    thumbnailUrl: PropTypes.string,
+  }).isRequired,
+  isLoading: PropTypes.bool.isRequired,
+  isError: PropTypes.bool.isRequired,
+  mapProps: PropTypes.shape({
+    basemap: PropTypes.string.isRequired,
+    labels: PropTypes.string.isRequired,
+    boundaries: PropTypes.bool.isRequired,
+    bounds: PropTypes.shape({}).isRequired,
+  }).isRequired,
+  activeLayers: PropTypes.arrayOf(
+    PropTypes.shape({}),
+  ).isRequired,
+  layerGroups: PropTypes.arrayOf(
+    PropTypes.shape({}),
+  ).isRequired,
+  activeInteractiveLayers: PropTypes.arrayOf(
+    PropTypes.shape({}),
+  ).isRequired,
+  user: PropTypes.shape({
+    id: PropTypes.string,
+    token: PropTypes.string,
+  }).isRequired,
+  webshot: PropTypes.bool.isRequired,
+};
+
+export default LayoutEmbedMap;